new update on the internal navbar component
